\name{Example_data}
\docType{data}
\alias{Example_data}
\title{Randomly Generated Data for Learning multipleComparisons}
\description{
  Dataframe including randomly generated data for two seasons (Dry and Wet), four plot locations (A, B, C, and D) four seed species (Spp_1, Spp_2, Spp_3, and Spp_4), and seed removal percentages. 
}
\usage{data(Example_data)}
\format{A \code{data.frame} object with data in the \code{Season},
  \code{Plot}, \code{Seed_species}, and \code{Seed_removal_percent} columns.  This is saved as an .RData file.}
\source{randomly generated data data}