\name{myCleanData}
\docType{data}
\alias{myCleanData}
\title{Example List Output from \code{DataCleaner}}
\description{
  List output from \code{DataCleaner} containing a list of lists and a dataframe.
}
\usage{data(myCleanData)}
\format{A list consisting of \code{compare_list} and \code{data_frame} generated by \code{DataCleaner} with \code{Example_data} to use in other functions in the \code{multipleComparisons} package.}
\source{output of \code{DataCleaner}}
\seealso{\code{\link{DataCleaner}}}
\keyword{datasets}