\name{MyCleanSubset}
\docType{data}
\alias{MyCleanSubset}
\title{Example List Output from \code{SubsetCleaner}}
\description{
  List output from \code{SubsetCleaner} containing a list of lists and a dataframe.
}
\usage{data(MyCleanSubset)}
\format{A list consisting of \code{sub_compare_list} and \code{sub_data_frame} generated by \code{SubsetCleaner} with \code{Example_data} to use in other functions in the \code{multipleComparisons} package.}
\source{output of \code{SubsetCleaner}}
\seealso{\code{\link{SubsetCleaner}}}
\keyword{datasets}